Main activity interface with the graph updated
