+++ android/app/src/main/java/com/example/hypnosapp/mainpage/Pantalla_Principal.java
@@ -80,103 +80,6 @@ public class Pantalla_Principal extends AppCompatActivity {
             tab.select();
         }
 
-        LineChart lineChart = findViewById(R.id.imagenGraficaSueñoDiario);
-
-        // Configura la gráfica
-        lineChart.getDescription().setEnabled(false);
-        lineChart.setTouchEnabled(true);
-        lineChart.setDragEnabled(true);
-        lineChart.setScaleEnabled(true);
-        lineChart.setPinchZoom(true);
-
-        FirebaseFirestore db = FirebaseFirestore.getInstance();
-        CollectionReference docRef = db.collection("users").document("R279SubMuPfIJf608GXGWbFOoTC2").collection("nightsData");
-
-        docRef.get().addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
-            @Override
-            public void onComplete(@NonNull Task<QuerySnapshot> task) {
-                if (task.isSuccessful()) {
-                    ArrayList<Entry> scores = new ArrayList<>();
-                    ArrayList<String> daysOfWeek = new ArrayList<>(); // To store the days of the week
-
-                    // Parse and store the data
-                    for (QueryDocumentSnapshot document : task.getResult()) {
-                        // Check if 'date' field exists and is of type Timestamp
-                        if (document.contains("date") && document.get("date") instanceof Timestamp) {
-                            Timestamp timestamp = (Timestamp) document.get("date");
-                            String date = formatDate(timestamp.toDate()); // Convert timestamp to Date and then to formatted string
-                            String dayOfWeek = getDayOfWeekFromDate(date);
-                            Float score = document.getDouble("score").floatValue();
-                            scores.add(new Entry(getTimestampFromDate(date), score != null ? score : 0f));
-                            daysOfWeek.add(dayOfWeek);
-                        } else {
-                            // Log a warning or handle the case where 'date' is not a Timestamp
-                            Log.w(TAG, "Invalid 'date' field in Firestore document");
-                        }
-                    }
-
-                    // Sort the entries based on the timestamp
-                    Collections.sort(scores, new Comparator<Entry>() {
-                        @Override
-                        public int compare(Entry entry1, Entry entry2) {
-                            return Float.compare(entry1.getX(), entry2.getX());
-                        }
-                    });
-
-// Create and configure the DataSet
-                    LineDataSet set = new LineDataSet(scores, "Puntuación de sueño");
-                    set.setFillAlpha(110);
-                    set.setColor(Color.parseColor("#164499"));
-                    set.setLineWidth(5f);
-                    set.setValueTextSize(15f);
-
-// Add the DataSet to the chart
-                    ArrayList<ILineDataSet> dataSets = new ArrayList<>();
-                    dataSets.add(set);
-                    LineData data = new LineData(dataSets);
-                    lineChart.setData(data);
-
-// Configure the appearance of the x-axis
-                    XAxis xAxis = lineChart.getXAxis();
-                    xAxis.setGranularity(1f);
-                    xAxis.setPosition(XAxis.XAxisPosition.BOTTOM);
-                    xAxis.setLabelRotationAngle(0f);
-
-// Manually set X-axis labels using timestamps
-                    xAxis.setValueFormatter(new ValueFormatter() {
-                        @Override
-                        public String getAxisLabel(float value, AxisBase axis) {
-                            long timestamp = (long) value;
-                            String date = formatDate(new Date(timestamp));
-                            return getDayOfWeekFromDate(date);
-                        }
-                    });
-
-// Set the number of labels to match the number of data points
-                    xAxis.setLabelCount(scores.size(), true);
-
-// Hide Y-axis labels and grid lines
-                    YAxis leftYAxis = lineChart.getAxisLeft();
-                    leftYAxis.setDrawLabels(false);
-                    leftYAxis.setDrawGridLines(false);
-
-                    YAxis rightYAxis = lineChart.getAxisRight();
-                    rightYAxis.setDrawLabels(false);
-                    rightYAxis.setDrawGridLines(false);
-
-// Hide X-axis grid lines
-                    xAxis.setDrawGridLines(false);
-
-// Invalidate the chart to refresh the appearance
-                    lineChart.invalidate();
-
-                } else {
-                    Log.d(TAG, "Error getting documents: ", task.getException());
-                }
-            }
-        });
-
-
         // FUNCIONALIDAD BOTONES MENUS
         MenuManager funcionMenu = new MenuManager();
 
@@ -244,42 +147,6 @@ public class Pantalla_Principal extends AppCompatActivity {
             }
         });
 
-    }
-
-    // Function to get the day of the week from the date string
-    private String getDayOfWeekFromDate(String date) {
-        SimpleDateFormat format = new SimpleDateFormat("dd 'de' MMMM 'de' yyyy, h:mm:ss a z", Locale.getDefault());
-        try {
-            Date parsedDate = format.parse(date);
-            Calendar calendar = Calendar.getInstance();
-            calendar.setTime(parsedDate);
-
-            // Use a different format for the day of the week
-            SimpleDateFormat dayFormat = new SimpleDateFormat("EEEE", Locale.getDefault());
-            return dayFormat.format(calendar.getTime());
-        } catch (ParseException e) {
-            e.printStackTrace();
-        }
-        return "";
-    }
-
-
-    // Function to get a timestamp from the date string
-    private float getTimestampFromDate(String date) {
-        SimpleDateFormat format = new SimpleDateFormat("dd 'de' MMMM 'de' yyyy, h:mm:ss a z", Locale.getDefault());
-        try {
-            Date parsedDate = format.parse(date);
-            return parsedDate.getTime(); // Convert to timestamp in milliseconds
-        } catch (ParseException e) {
-            e.printStackTrace();
-        }
-        return 0f;
-    }
-
-    // Function to format Date to a string
-    private String formatDate(Date date) {
-        SimpleDateFormat format = new SimpleDateFormat("dd 'de' MMMM 'de' yyyy, h:mm:ss a z", Locale.getDefault());
-        return format.format(date);
     }
 
 }
